+++ test/test_autograd.py
@@ -45,11 +45,11 @@ from torch.testing._internal.common_device_type import (instantiate_device_type_
                                                         deviceCountAtLeast, skipCUDAIfCudnnVersionLessThan,
                                                         skipCUDAIf, skipMeta)
 
-from autograd.test_complex import TestAutogradComplex
+import autograds.test_complex
 
 load_tests = make_load_tests(
     imported_test_cases={
-        TestAutogradComplex,
+        autograds.test_complex.TestAutogradComplex,
     })
 
 import pickle
